+++ npm/vue/src/index.ts
@@ -21,10 +21,9 @@ import type {
 } from 'vue'
 import type { MountingOptions as VTUMountingOptions, VueWrapper } from '@vue/test-utils'
 import {
-  injectStylesBeforeElement,
-  StyleOptions,
   getContainerEl,
   setupHooks,
+  checkForRemovedStyleOptions,
 } from '@cypress/mount-utils'
 
 import * as _VueTestUtils from '@vue/test-utils'
@@ -44,6 +43,7 @@ const {
 export { VueTestUtils }
 
 const DEFAULT_COMP_NAME = 'unknown'
+const VUE_ROOT = '__cy_vue_root'
 
 type GlobalMountOptions = Required<VTUMountingOptions<any>>['global']
 
@@ -68,28 +68,18 @@ type MountingOptions<Props, Data = {}> = Omit<VTUMountingOptions<Props, Data>, '
     use?: GlobalMountOptions['plugins']
     mixin?: GlobalMountOptions['mixins']
   }
-} & Partial<StyleOptions>
+}
 
 export type CyMountOptions<Props, Data = {}> = MountingOptions<Props, Data>
 
-Cypress.on('run:start', () => {
-  // `mount` is designed to work with component testing only.
-  // it assumes ROOT_SELECTOR exists, which is not the case in e2e.
-  // if the user registers a custom command that imports `cypress/vue`,
-  // this event will be registered and cause an error when the user
-  // launches e2e (since it's common to use Cypress for both CT and E2E.
-  // https://github.com/cypress-io/cypress/issues/17438
-  if (Cypress.testingType !== 'component') {
-    return
-  }
+const cleanup = () => {
+  Cypress.vueWrapper?.unmount()
+  Cypress.$(`#${VUE_ROOT}`).remove()
 
-  Cypress.on('test:before:run', () => {
-    Cypress.vueWrapper?.unmount()
-    const el = getContainerEl()
+  ;(Cypress as any).vueWrapper = null
 
-    el.innerHTML = ''
-  })
-})
+  ;(Cypress as any).vue = null
+}
 
 /**
  * The types for mount have been copied directly from the VTU mount
@@ -129,7 +119,10 @@ export function mount<V extends {}>(
     __vccOpts: any
   },
   options?: MountingOptions<any> & Record<string, any>
-): Cypress.Chainable<VueWrapper<ComponentPublicInstance<V>>>
+): Cypress.Chainable<{
+  wrapper: VueWrapper<ComponentPublicInstance<V>>
+  component: VueWrapper<ComponentPublicInstance<V>>['vm']
+}>
 
 // Class component (without vue-class-component) - props
 export function mount<V extends {}, P>(
@@ -139,7 +132,10 @@ export function mount<V extends {}, P>(
     defaultProps?: Record<string, Prop<any>> | string[]
   },
   options?: MountingOptions<P & PublicProps> & Record<string, any>
-): Cypress.Chainable<VueWrapper<ComponentPublicInstance<V>>>
+): Cypress.Chainable<{
+  wrapper: VueWrapper<ComponentPublicInstance<V>>
+  component: VueWrapper<ComponentPublicInstance<V>>['vm']
+}>
 
 // Class component - no props
 export function mount<V extends {}>(
@@ -148,7 +144,10 @@ export function mount<V extends {}>(
     registerHooks(keys: string[]): void
   },
   options?: MountingOptions<any> & Record<string, any>
-): Cypress.Chainable<VueWrapper<ComponentPublicInstance<V>>>
+): Cypress.Chainable<{
+  wrapper: VueWrapper<ComponentPublicInstance<V>>
+  component: VueWrapper<ComponentPublicInstance<V>>['vm']
+}>
 
 // Class component - props
 export function mount<V extends {}, P>(
@@ -158,13 +157,19 @@ export function mount<V extends {}, P>(
     registerHooks(keys: string[]): void
   },
   options?: MountingOptions<P & PublicProps> & Record<string, any>
-): Cypress.Chainable<VueWrapper<ComponentPublicInstance<V>>>
+): Cypress.Chainable<{
+  wrapper: VueWrapper<ComponentPublicInstance<V>>
+  component: VueWrapper<ComponentPublicInstance<V>>['vm']
+}>
 
 // Functional component with emits
 export function mount<Props extends {}, E extends EmitsOptions = {}>(
   originalComponent: FunctionalComponent<Props, E>,
   options?: MountingOptions<Props & PublicProps> & Record<string, any>
-): Cypress.Chainable<VueWrapper<ComponentPublicInstance<Props>>>
+): Cypress.Chainable<{
+  wrapper: VueWrapper<ComponentPublicInstance<Props>>
+  component: VueWrapper<ComponentPublicInstance<Props>>['vm']
+}>
 
 // Component declared with defineComponent
 export function mount<
@@ -200,8 +205,8 @@ export function mount<
     D
   > &
   Record<string, any>
-): Cypress.Chainable<
-  VueWrapper<
+): Cypress.Chainable<{
+  wrapper: VueWrapper<
     InstanceType<
       DefineComponent<
         PropsOrPropOptions,
@@ -219,13 +224,34 @@ export function mount<
       >
     >
   >
+  component: VueWrapper<
+    InstanceType<
+      DefineComponent<
+        PropsOrPropOptions,
+        RawBindings,
+        D,
+        C,
+        M,
+        Mixin,
+        Extends,
+        E,
+        EE,
+        PP,
+        Props,
+        Defaults
+      >
+    >
+  >['vm']}
 >
 
 // component declared by vue-tsc ScriptSetup
 export function mount<T extends DefineComponent<any, any, any, any>>(
   component: T,
   options?: ComponentMountingOptions<T>
-): Cypress.Chainable<VueWrapper<InstanceType<T>>>
+): Cypress.Chainable<{
+  wrapper: VueWrapper<InstanceType<T>>
+  component: VueWrapper<InstanceType<T>>['vm']
+}>
 
 // Component declared with no props
 export function mount<
@@ -251,7 +277,10 @@ export function mount<
     EE
   >,
   options?: MountingOptions<Props & PublicProps, D>
-): Cypress.Chainable<VueWrapper<ComponentPublicInstance<Props, RawBindings, D, C, M, E, VNodeProps & Props>>> & Record<string, any>
+): Cypress.Chainable<{
+  wrapper: VueWrapper<ComponentPublicInstance<Props, RawBindings, D, C, M, E, VNodeProps & Props>>
+  component: VueWrapper<ComponentPublicInstance<Props, RawBindings, D, C, M, E, VNodeProps & Props>>['vm']
+}> & Record<string, any>
 
 // Component declared with { props: [] }
 export function mount<
@@ -281,7 +310,10 @@ export function mount<
     Props
   >,
   options?: MountingOptions<Props & PublicProps, D>
-): Cypress.Chainable<VueWrapper<ComponentPublicInstance<Props, RawBindings, D, C, M, E>>>
+): Cypress.Chainable<{
+  wrapper: VueWrapper<ComponentPublicInstance<Props, RawBindings, D, C, M, E>>
+  component: VueWrapper<ComponentPublicInstance<Props, RawBindings, D, C, M, E>>['vm']
+}>
 
 // Component declared with { props: { ... } }
 export function mount<
@@ -309,8 +341,8 @@ export function mount<
     EE
   >,
   options?: MountingOptions<ExtractPropTypes<PropsOptions> & PublicProps, D>
-): Cypress.Chainable<
-  VueWrapper<
+): Cypress.Chainable<{
+  wrapper: VueWrapper<
     ComponentPublicInstance<
       ExtractPropTypes<PropsOptions>,
       RawBindings,
@@ -321,18 +353,46 @@ export function mount<
       VNodeProps & ExtractPropTypes<PropsOptions>
     >
   >
->
+  component: VueWrapper<
+    ComponentPublicInstance<
+      ExtractPropTypes<PropsOptions>,
+      RawBindings,
+      D,
+      C,
+      M,
+      E,
+      VNodeProps & ExtractPropTypes<PropsOptions>
+    >
+  >['vm']
+}>
 
-// implementation
+/**
+ * Mounts a component and returns an object containing the component and VueWrapper
+ * @param componentOptions
+ * @param options
+ * @returns {Cypress.Chainable<{wrapper: VueWrapper<T>, component: T}>}
+ * @see {@link https://on.cypress.io/mounting-vue} for more details.
+ * @example
+ * import { mount } from '@cypress/vue'
+ * import { Stepper } from './Stepper.vue'
+ *
+ * it('mounts', () => {
+ *   cy.mount(Stepper)
+ *   cy.get('[data-cy=increment]').click()
+ *   cy.get('[data-cy=counter]').should('have.text', '1')
+ * })
+ */
 export function mount (componentOptions: any, options: any = {}) {
-  // wait for cypress to load
+  checkForRemovedStyleOptions(options)
+  // Remove last mounted component if cy.mount is called more than once in a test
+  cleanup()
+
+  // then wait for cypress to load
   return cy.then(() => {
     // @ts-ignore
     const document: Document = cy.state('document')
 
     const el = getContainerEl()
-
-    injectStylesBeforeElement(options, document, el)
 
     // merge the extensions with global
     if (options.extensions) {
@@ -343,7 +403,7 @@ export function mount (componentOptions: any, options: any = {}) {
 
     const componentNode = document.createElement('div')
 
-    componentNode.id = '__cy_vue_root'
+    componentNode.id = VUE_ROOT
 
     el.append(componentNode)
 
@@ -354,7 +414,6 @@ export function mount (componentOptions: any, options: any = {}) {
     Cypress.vue = wrapper.vm as ComponentPublicInstance
 
     return cy
-    .wrap(wrapper, { log: false })
     .wait(1, { log: false })
     .then(() => {
       if (options.log !== false) {
@@ -367,9 +426,22 @@ export function mount (componentOptions: any, options: any = {}) {
         })
       }
 
-      // by returning undefined we keep the previous subject
-      // which is the mounted component
-      return undefined
+      const returnVal = {
+        wrapper,
+        component: wrapper.vm,
+      }
+
+      return new Proxy(Object.create(returnVal), {
+        get (obj, prop) {
+        // throw an error if it looks like the caller is trying to call a method on the VueWrapper that was originally returned
+          if (Reflect.get(wrapper, prop)) {
+            // @ts-expect-error - internal API
+            Cypress.utils.throwErrByPath('mount.vue_yielded_value')
+          }
+
+          return Reflect.get(obj, prop)
+        },
+      })
     })
   })
 }
@@ -403,13 +475,17 @@ function getComponentDisplayName (componentOptions: any): string {
  * @example
  *  import {mountCallback} from '@cypress/vue'
  *  beforeEach(mountVue(component, options))
+ *
+ * Removed as of Cypress 11.0.0.
+ * @see https://on.cypress.io/migration-11-0-0-component-testing-updates
  */
 export function mountCallback (
   component: any,
   options: any = {},
-): () => Cypress.Chainable {
+) {
   return () => {
-    return mount(component, options)
+    // @ts-expect-error - undocumented API
+    Cypress.utils.throwErrByPath('mount.mount_callback')
   }
 }
 
@@ -421,4 +497,4 @@ export function mountCallback (
 //    import { registerCT } from 'cypress/<my-framework>'
 //    registerCT()
 // Note: This would be a breaking change
-setupHooks()
+setupHooks(cleanup)