+++ packages/proxy/lib/http/request-middleware.ts
@@ -156,8 +156,9 @@ const MaybeEndRequestWithBufferedResponse: RequestMiddleware = function () {
 
   if (buffer) {
     this.debug('ending request with buffered response')
-    // NOTE: Only inject fullCrossOrigin here if the super domain origins do not match in order to keep parity with cypress application reloads
-    this.res.wantsInjection = buffer.isCrossSuperDomainOrigin ? 'fullCrossOrigin' : 'full'
+    // NOTE: Only inject fullCrossOrigin here if experimental is on and
+    // the super domain origins do not match in order to keep parity with cypress application reloads
+    this.res.wantsInjection = buffer.urlDoesNotMatchPolicyBasedOnDomain ? 'fullCrossOrigin' : 'full'
 
     return this.onResponse(buffer.response, buffer.stream)
   }