+++ system-tests/projects/session-and-origin-e2e-specs/cypress/e2e/session/recreates_session.cy.js
@@ -8,7 +8,7 @@ before(() => {
   setupFn = cy.stub().as('runSetup')
   validateFn = cy.stub().callsFake(() => {
     if (validateFn.callCount === 2) {
-      return false
+      expect(true).to.be.false
     }
   }).as('runValidation')
 })